server/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
langchain==0.2.1
langchain_chroma==0.1.1
langchain_community==0.2.1
langchain_core==0.2.1
langchain_openai==0.1.7
langchain_text_splitters==0.2.0
python-dotenv==1.0.1
